+++ pj_plugins/tests/file_source_integration_test.cpp
@@ -63,7 +63,7 @@ bool whEnsureField(
   return true;
 }
 bool whAppendRecord(
-    void* ctx, PJ_topic_handle_t, int64_t, const PJ_named_field_value_t*, size_t, PJ_error_t* err) noexcept {
+    void* ctx, PJ_topic_handle_t, int64_t, const PJ_named_field_value_t*, uint64_t, PJ_error_t* err) noexcept {
   auto* s = static_cast<WriteHostState*>(ctx);
   if (s->fail_next_append) {
     s->last_error = "mock append failure";
@@ -73,7 +73,7 @@ bool whAppendRecord(
   return true;
 }
 bool whAppendBoundRecord(
-    void*, PJ_topic_handle_t, int64_t, const PJ_bound_field_value_t*, size_t, PJ_error_t*) noexcept {
+    void*, PJ_topic_handle_t, int64_t, const PJ_bound_field_value_t*, uint64_t, PJ_error_t*) noexcept {
   return true;
 }
 bool whAppendArrowStream(
